@@ -5,15 +5,15 @@
   "scripts": {
     "lint": "eslint \"{src,tests}/**/*.{js,ts}\" --max-warnings=0",
     "lint:fix": "npm run lint -- --fix",
-    "build:localhost": "vue-cli-service build --mode localhost",
-    "build:production": "vue-cli-service build --mode prod",
-    "build:staging": "vue-cli-service build --mode staging",
-    "start": "vue-cli-service serve --mode localhost --port 8081",
-    "start:dev": "vue-cli-service serve --mode localhost --port 8081",
-    "start:dev:local": "set -a && . ./.env.dist.local && . ./.env.override.local && set +a && vue-cli-service serve --mode localhost --port 8081",
+    "build:localhost": "vite build --mode localhost",
+    "build:production": "vite build --mode prod",
+    "build:staging": "vite build --mode staging",
+    "start": "vite --host",
+    "start:dev": "vite --host",
+    "start:dev:local": "set -a && . ./.env.dist.local && . ./.env.override.local && set +a && vite --mode localhost --port 8081",
     "test": "node tests/localhost-runner.js",
-    "test:e2e:open": "wait-on http://localhost:8081 && npx cypress open",
-    "test:e2e:run": "wait-on http://localhost:8081 && npx cypress run --headless --browser chrome",
+    "test:e2e:open": "npx cypress open",
+    "test:e2e:run": "npx cypress run --headed --browser chrome",
     "test:e2e:staging:open": "npx cypress open --config-file cypress.staging.config.js",
     "test:e2e:staging:run": "npx cypress run --config-file cypress.staging.config.js",
     "tailwind:view": "tailwind-config-viewer -o"
@@ -29,6 +29,7 @@
     "@tiptap/extension-placeholder": "^2.0.0-beta.199",
     "@tiptap/starter-kit": "^2.0.0-beta.199",
     "@tiptap/vue-3": "^2.0.0-beta.199",
+    "@vitejs/plugin-vue": "^4.2.3",
     "@vue/eslint-config-airbnb": "^7.0.0",
     "@vuelidate/core": "^2.0.1",
     "@vuelidate/validators": "^2.0.1",
@@ -38,7 +39,6 @@
     "chartjs-adapter-moment": "^1.0.1",
     "chartjs-plugin-annotation": "^2.2.1",
     "chartkick": "github:CrowdDotDev/chartkick.js",
-    "core-js": "3.24.1",
     "element-plus": "^2.3.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.27.5",
@@ -59,7 +59,8 @@
     "unleash-proxy-client": "^2.4.3",
     "uuid": "8.3.0",
     "v-network-graph": "^0.8.2",
-    "vue": "^3.2.45",
+    "vite": "^4.3.9",
+    "vue": "^3.3.1",
     "vue-3-sanitize": "^0.1.4",
     "vue-grid-layout": "3.0.0-beta1",
     "vue-json-pretty": "^2.2.4",
@@ -68,15 +69,14 @@
     "vue3-lazyload": "^0.3.6",
     "vuedraggable": "^4.1.0",
     "vuex": "^4.1.0",
-    "webpack": "^5.76.3",
     "yup": "^0.32.11"
   },
   "devDependencies": {
     "@tailwindcss/line-clamp": "^0.4.2",
-    "@vue/cli-plugin-babel": "^5.0.8",
-    "@vue/cli-plugin-eslint": "^5.0.8",
-    "@vue/cli-service": "^5.0.8",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "@vue/compiler-sfc": "^3.2.29",
+    "@vue/eslint-config-typescript": "^11.0.3",
     "autoprefixer": "^10.4.14",
     "concurrently": "^7.6.0",
     "css-minimizer-webpack-plugin": "^4.2.2",
@@ -88,15 +88,14 @@
     "eslint-plugin-vue": "^9.10.0",
     "mini-css-extract-plugin": "^2.7.5",
     "postcss": "^8.4.21",
+    "rollup-plugin-visualizer": "^5.9.2",
     "sass": "^1.60.0",
-    "sass-loader": "^12.6.0",
     "tailwind-config-viewer": "^1.7.2",
     "tailwindcss": "^3.2.7",
     "terser-webpack-plugin": "^5.3.7",
+    "typescript": "~5.0.4",
     "unplugin-auto-import": "^0.15.2",
     "unplugin-vue-components": "^0.24.1",
-    "wait-on": "^6.0.1",
-    "webpack-bundle-analyzer": "^4.8.0",
-    "webpack-deadcode-plugin": "^0.1.17"
+    "wait-on": "^6.0.1"
   }
 }
