@@ -27,7 +27,7 @@
             <router-link
               v-if="
                 hasPermissionToCreate
-                  && (hasIntegrations || hasMembers)
+                  && (hasIntegrations || membersCount > 0)
               "
               :to="{
                 name: 'memberCreate',
@@ -51,175 +51,122 @@
         </div>
       </div>
 
-      <app-member-list-tabs />
-      <app-member-list-filter
-        v-if="hasMembers"
+      <cr-saved-views v-model="filters" :config="memberSavedViews" :views="memberViews" />
+      <cr-filter
+        v-if="customAttributesFilter"
+        v-model="filters"
+        :config="memberFilters"
+        :search-config="memberSearchFilter"
+        :saved-views-config="memberSavedViews"
+        :custom-config="customAttributesFilter"
+        @fetch="fetch($event)"
       />
       <app-member-list-table
         :has-integrations="hasIntegrations"
-        :has-members="hasMembers"
-        :is-page-loading="isPageLoading"
+        :has-members="membersCount > 0"
+        :is-page-loading="loading"
       />
     </div>
   </app-page-wrapper>
 </template>
 
-<script>
-import { mapGetters, mapActions, mapMutations } from 'vuex';
-import moment from 'moment';
-import MemberListFilter from '@/modules/member/components/list/member-list-filter.vue';
-import MemberListTable from '@/modules/member/components/list/member-list-table.vue';
-import MemberListTabs from '@/modules/member/components/list/member-list-tabs.vue';
-import PageWrapper from '@/shared/layout/page-wrapper.vue';
-import { MemberService } from '../member-service';
-import { MemberPermissions } from '../member-permissions';
+<script setup lang="ts">
+import AppPageWrapper from '@/shared/layout/page-wrapper.vue';
+import CrFilter from '@/shared/modules/filters/components/Filter.vue';
+import { useMemberStore } from '@/modules/member/store/pinia';
+import { storeToRefs } from 'pinia';
+import { ref, onMounted, computed } from 'vue';
+import { MemberService } from '@/modules/member/member-service';
+import { MemberPermissions } from '@/modules/member/member-permissions';
+import { mapGetters } from '@/shared/vuex/vuex.helpers';
+import { FilterQuery } from '@/shared/modules/filters/types/FilterQuery';
+import CrSavedViews from '@/shared/modules/saved-views/components/SavedViews.vue';
+import AppMemberListTable from '@/modules/member/components/list/member-list-table.vue';
+import { memberFilters, memberSearchFilter } from '../config/filters/main';
+import { memberSavedViews, memberViews } from '../config/saved-views/main';
 
-export default {
-  name: 'AppMemberListPage',
+const memberStore = useMemberStore();
+const { getMemberCustomAttributes, fetchMembers } = memberStore;
+const { filters, customAttributesFilter, savedFilterBody } = storeToRefs(memberStore);
 
-  components: {
-    'app-member-list-filter': MemberListFilter,
-    'app-member-list-table': MemberListTable,
-    'app-member-list-tabs': MemberListTabs,
-    'app-page-wrapper': PageWrapper,
-  },
+const membersCount = ref(0);
+const membersToMergeCount = ref(0);
 
-  data() {
-    return {
-      membersToMergeCount: 0,
-      hasMembers: false,
-      isPageLoading: true,
-    };
-  },
+const { listByPlatform } = mapGetters('integration');
+const { currentUser, currentTenant } = mapGetters('auth');
 
-  computed: {
-    ...mapGetters({
-      currentUser: 'auth/currentUser',
-      currentTenant: 'auth/currentTenant',
-      integrations: 'integration/listByPlatform',
-      activeView: 'member/activeView',
-    }),
+const hasIntegrations = computed(() => !!Object.keys(listByPlatform.value || {}).length);
 
-    hasIntegrations() {
-      return !!Object.keys(this.integrations || {}).length;
-    },
+const hasPermissionToCreate = computed(() => new MemberPermissions(
+  currentTenant.value,
+  currentUser.value,
+)?.create);
 
-    hasPermissionToCreate() {
-      return new MemberPermissions(
-        this.currentTenant,
-        this.currentUser,
-      ).create;
-    },
+const isCreateLockedForSampleData = computed(() => new MemberPermissions(
+  currentTenant.value,
+  currentUser.value,
+)?.createLockedForSampleData);
 
-    isCreateLockedForSampleData() {
-      return new MemberPermissions(
-        this.currentTenant,
-        this.currentUser,
-      ).createLockedForSampleData;
-    },
+const isEditLockedForSampleData = computed(() => new MemberPermissions(
+  currentTenant.value,
+  currentUser.value,
+)?.editLockedForSampleData);
 
-    isEditLockedForSampleData() {
-      return new MemberPermissions(
-        this.currentTenant,
-        this.currentUser,
-      ).editLockedForSampleData;
-    },
-  },
-
-  async created() {
-    const { filter } = this.$store.state.member;
-
-    this.isPageLoading = true;
-
-    await this.doFetchIntegrations();
-    await this.doFetchCustomAttributes();
-
-    const { joinedFrom, activeFrom } = this.$route.query;
-    if (
-      joinedFrom
-      && moment(joinedFrom, 'YYYY-MM-DD', true).isValid()
-    ) {
-      await this.updateFilterAttribute({
-        custom: false,
-        defaultOperator: 'gt',
-        defaultValue: joinedFrom,
-        expanded: false,
-        label: 'Joined date',
-        name: 'joinedAt',
-        operator: 'gt',
-        type: 'date',
-        value: joinedFrom,
-      });
-    }
-    if (
-      activeFrom
-      && moment(activeFrom, 'YYYY-MM-DD', true).isValid()
-    ) {
-      await this.updateFilterAttribute({
-        custom: false,
-        defaultOperator: 'eq',
-        defaultValue: activeFrom,
-        expanded: false,
-        label: 'Last activity date',
-        name: 'lastActive',
-        operator: 'gt',
-        type: 'date',
-        value: activeFrom,
-      });
-      this.SORTER_CHANGED({
-        activeView: this.activeView,
-        sorter: {
-          prop: 'activityCount',
-          order: 'descending',
-        },
-      });
-    }
-    await this.doFetch({
-      filter,
-      keepPagination: true,
+const fetchMembersToMergeCount = () => {
+  MemberService.fetchMergeSuggestions(1, 0)
+    .then(({ count }: any) => {
+      membersToMergeCount.value = count;
     });
-
-    const membersList = await this.doGetMembersCount();
-    const mergeSuggestions = await MemberService.fetchMergeSuggestions(1, 0);
-
-    this.membersToMergeCount = mergeSuggestions.count;
-    this.hasMembers = membersList.count > 0;
-    this.isPageLoading = false;
-  },
-
-  async mounted() {
-    window.analytics.page('Members');
-  },
-
-  methods: {
-    ...mapMutations({
-      SORTER_CHANGED: 'member/SORTER_CHANGED',
-    }),
-    ...mapActions({
-      doFetchWidgets: 'widget/doFetch',
-      doFetchCustomAttributes:
-        'member/doFetchCustomAttributes',
-      updateFilterAttribute: 'member/updateFilterAttribute',
-      doFetch: 'member/doFetch',
-      doFetchIntegrations: 'integration/doFetch',
-    }),
-
-    async doGetMembersCount() {
-      try {
-        const response = await MemberService.list(
-          {},
-          '',
-          1,
-          0,
-          undefined,
-          true,
-        );
-
-        return response;
-      } catch (e) {
-        return null;
-      }
-    },
-  },
 };
+
+const loading = ref(true);
+
+const doGetMembersCount = () => {
+  (MemberService.listMembers({
+    limit: 1,
+    offset: 0,
+  }, true) as Promise<any>)
+    .then(({ count }) => {
+      membersCount.value = count;
+    });
+};
+
+const showLoading = (filter: any, body: any): boolean => {
+  const saved: any = { ...savedFilterBody.value };
+  delete saved.offset;
+  delete saved.limit;
+  delete saved.orderBy;
+  const compare = {
+    ...body,
+    filter,
+  };
+  return JSON.stringify(saved) !== JSON.stringify(compare);
+};
+
+const fetch = ({
+  filter, offset, limit, orderBy, body,
+}: FilterQuery) => {
+  if (!loading.value) {
+    loading.value = showLoading(filter, body);
+  }
+  fetchMembers({
+    body: {
+      ...body,
+      filter,
+      offset,
+      limit,
+      orderBy,
+    },
+  })
+    .finally(() => {
+      loading.value = false;
+    });
+};
+
+onMounted(() => {
+  fetchMembersToMergeCount();
+  doGetMembersCount();
+  getMemberCustomAttributes();
+  (window as any).analytics.page('Members');
+});
 </script>
