@@ -3,15 +3,18 @@ module.exports = {
 
   env: {
     node: true,
+    es2022: true,
   },
 
   extends: [
     'plugin:vue/vue3-recommended',
     '@vue/airbnb',
+    '@vue/typescript',
   ],
 
   parserOptions: {
     ecmaVersion: 2020,
+    parser: '@typescript-eslint/parser',
   },
 
   rules: {
@@ -37,11 +40,13 @@ module.exports = {
     'keyword-spacing': 'warn',
     'space-before-blocks': 'warn',
     quotes: 'warn',
-    'no-unused-vars': 'warn',
+    'no-unused-vars': 'off',
     'no-multiple-empty-lines': 'warn',
     'vue/no-v-html': 'off',
     'import/prefer-default-export': 'off',
+    'import/no-named-as-default': 'off',
     'class-methods-use-this': 'off',
+    'no-shadow': 'off',
     'vuejs-accessibility/mouse-events-have-key-events': 'off',
     'vuejs-accessibility/click-events-have-key-events': 'off',
     'func-names': 'off',
@@ -65,6 +70,7 @@ module.exports = {
       },
     ],
   },
+
   settings: {
     'import/resolver': {
       alias: {
@@ -73,6 +79,7 @@ module.exports = {
           ['@', './src'],
         ],
       },
+      typescript: {},
     },
   },
 };
