@@ -1,5 +1,4 @@
 const { defineConfig } = require('cypress');
-const webpackConfig = require('@vue/cli-service/webpack.config.js');
 require('dotenv').config({
   path: './.env.cypress',
 });
@@ -15,15 +14,17 @@ module.exports = defineConfig({
     screenshotsFolder: 'tests/screenshots',
     videosFolder: 'tests/videos',
   },
+  numTestsKeptInMemory: 1,
   component: {
     devServer: {
       framework: 'vue',
-      bundler: 'webpack',
-      webpackConfig,
+      bundler: 'vite',
     },
   },
   browser: {
     chromeWebSecurity: false,
+    numTestsKeptInMemory: 1,
+    experimentalMemoryManagement: true,
   },
   env: {
     appUrl: 'http://localhost:8081',
