@@ -1,5 +1,5 @@
 import AuthGuard from '@/middleware/auth/auth-guard';
-// import UnauthGuard from '@/middleware/auth/unauth-guard';
+import UnauthGuard from '@/middleware/auth/unauth-guard';
 // import EmailAlreadyVerifiedGuard from '@/middleware/auth/email-already-verified-guard';
 // import PermissionGuard from '@/middleware/auth/permission-guard';
 // import NotEmptyTenant from '@/middleware/auth/not-empty-tenant-guard';
@@ -8,7 +8,7 @@ import AuthGuard from '@/middleware/auth/auth-guard';
 /* Temporarly disabling guards, only AuthGuard has been working before and other caused too much issues when enabled */
 export default [
   AuthGuard,
-  // UnauthGuard,
+  UnauthGuard,
   // EmailAlreadyVerifiedGuard,
   // PermissionGuard,
   // NotEmptyTenant,
