@@ -75,10 +75,14 @@
               <router-link
                 :to="{
                   name: 'organization',
-                  query: {
-                    activeTab: 'new-and-active',
-                    joinedFrom: periodStartDate,
-                  },
+                  query: filterQueryService().setQuery({
+                    ...newAndActive.filter,
+                    joinedDate: {
+                      include: true,
+                      value: periodStartDate,
+                      operator: 'gt',
+                    },
+                  }),
                 }"
                 class="text-sm leading-5 font-medium text-red"
               >
@@ -152,10 +156,14 @@
               <router-link
                 :to="{
                   name: 'organization',
-                  query: {
-                    activeTab: 'all',
-                    activeFrom: periodStartDate,
-                  },
+                  query: filterQueryService().setQuery({
+                    ...allOrganizations.filter,
+                    lastActivityDate: {
+                      include: true,
+                      value: periodStartDate,
+                      operator: 'gt',
+                    },
+                  }),
                 }"
                 class="text-sm leading-5 font-medium text-red"
               >
@@ -184,6 +192,9 @@ import { DAILY_GRANULARITY_FILTER } from '@/modules/widget/widget-constants';
 import AppDashboardEmptyState from '@/modules/dashboard/components/dashboard-empty-state.vue';
 import AppDashboardWidgetHeader from '@/modules/dashboard/components/dashboard-widget-header.vue';
 import AppDashboardWidgetChart from '@/modules/dashboard/components/dashboard-widget-chart.vue';
+import newAndActive from '@/modules/organization/config/saved-views/views/new-and-active';
+import allOrganizations from '@/modules/organization/config/saved-views/views/all-organizations';
+import { filterQueryService } from '@/shared/modules/filters/services/filter-query.service';
 
 export default {
   name: 'AppDashboardOrganizations',
@@ -200,6 +211,9 @@ export default {
       activeOrganizationChart,
       newOrganizationCount,
       activeOrganizationCount,
+      filterQueryService,
+      newAndActive,
+      allOrganizations,
     };
   },
   computed: {
