@@ -121,6 +121,8 @@ import ConfirmDialog from '@/shared/dialog/confirm-dialog';
 import getCustomAttributes from '@/shared/fields/get-custom-attributes';
 import getAttributesModel from '@/shared/attributes/get-attributes-model';
 import getParsedAttributes from '@/shared/attributes/get-parsed-attributes';
+import { useMemberStore } from '@/modules/member/store/pinia';
+import { storeToRefs } from 'pinia';
 
 const LoaderIcon = h(
   'i',
@@ -137,9 +139,13 @@ const ArrowPrevIcon = h(
   [],
 );
 
+const { getMemberCustomAttributes } = useMemberStore();
 const router = useRouter();
 const route = useRoute();
 const store = useStore();
+
+const memberStore = useMemberStore();
+const { customAttributes } = storeToRefs(memberStore);
 
 const { fields } = MemberModel;
 const formSchema = computed(
@@ -152,8 +158,7 @@ const formSchema = computed(
     fields.organizations,
     fields.attributes,
     ...getCustomAttributes({
-      customAttributes:
-      store.state.member.customAttributes,
+      customAttributes: customAttributes.value,
       considerShowProperty: false,
     }),
   ]),
@@ -212,7 +217,7 @@ const isDrawerOpen = ref(false);
 const rules = reactive(formSchema.value.rules());
 
 const computedFields = computed(() => fields);
-const computedAttributes = computed(() => Object.values(store.state.member.customAttributes));
+const computedAttributes = computed(() => Object.values(customAttributes.value));
 
 // UI Validations
 const isEditPage = computed(() => !!route.params.id);
@@ -249,7 +254,7 @@ onBeforeRouteLeave((to) => {
 
 onMounted(async () => {
   // Fetch custom attributes on mount
-  await store.dispatch('member/doFetchCustomAttributes');
+  await getMemberCustomAttributes();
 
   if (isEditPage.value) {
     const { id } = route.params;
