@@ -76,6 +76,8 @@ import AppMemberViewAside from '@/modules/member/components/view/member-view-asi
 import AppMemberViewNotes from '@/modules/member/components/view/member-view-notes.vue';
 import AppMemberViewContributions from '@/modules/member/components/view/member-view-contributions.vue';
 import AppMemberViewTasks from '@/modules/member/components/view/member-view-tasks.vue';
+import { useMemberStore } from '@/modules/member/store/pinia';
+import { storeToRefs } from 'pinia';
 
 const store = useStore();
 const props = defineProps({
@@ -86,6 +88,10 @@ const props = defineProps({
 });
 
 const { currentTenant, currentUser } = mapGetters('auth');
+
+const memberStore = useMemberStore();
+const { customAttributes } = storeToRefs(memberStore);
+const { getMemberCustomAttributes } = memberStore;
 
 const member = computed(() => store.getters['member/find'](props.id) || {});
 
@@ -109,11 +115,12 @@ const tab = ref('activities');
 
 onMounted(async () => {
   await store.dispatch('member/doFind', props.id);
+
   if (
-    Object.keys(store.state.member.customAttributes)
+    Object.keys(customAttributes.value)
       .length === 0
   ) {
-    await store.dispatch('member/doFetchCustomAttributes');
+    await getMemberCustomAttributes();
   }
   loading.value = false;
 });
