@@ -43,7 +43,7 @@ function selectErrorCode(error) {
 
 export default class Errors {
   static handle(error) {
-    if (process.env.NODE_ENV !== 'test') {
+    if (import.meta.env.NODE_ENV !== 'test') {
       console.error(selectErrorMessage(error));
       console.error(error);
     }
