@@ -35,7 +35,7 @@
 </template>
 
 <script setup>
-import { defineEmits, defineProps, computed } from 'vue';
+import { computed } from 'vue';
 
 const props = defineProps({
   value: {
