@@ -14,7 +14,7 @@ export default () => {
   });
 
   it('Filters by each custom attribute', () => {
-    cy.get('#filterList .custom-attribute').each((attribute) => {
+    cy.get('[data-qa="filter-list-item-custom"]').each((attribute) => {
       const attributeText = attribute.text().trim();
       const attributeName = attributeText.split(' ').map((text, textIndex) => {
         if (textIndex > 0) {
@@ -25,11 +25,14 @@ export default () => {
       cy.get('[data-qa="filter-dropdown"]').click();
       cy.wrap(attribute).click();
       cy.get('.filter-list-item-popper').then(($filter) => {
-        if ($filter.find('.filter-type-select-multi').length > 0) {
-          /** ***********\
-           * SELECT-MULTI
-           \************ */
-          cy.get('[data-qa="filter-select-option"]').each((option) => {
+        if ($filter.hasClass('filter-multiselect')) {
+          // /** ***********\
+          //  * SELECT-MULTI
+          //  \************ */
+          cy.get('[data-qa="filter-select-option"]').each((option, index) => {
+            if(index > 15){
+              return;
+            }
             const optionText = option.text().trim();
             cy.wrap(option).click();
             cy.get('[data-qa="filter-apply"]').click();
@@ -45,10 +48,9 @@ export default () => {
             cy.scrollTo(0, 0);
             cy.wait(200);
             cy.get('[data-qa="filter-list-chip"]').click({ force: true });
-            cy.get('[data-qa="filter-list-chip"]').click({ force: true });
             cy.wrap(option).click();
           });
-        } else if ($filter.find('.filter-type-boolean').length > 0) {
+        } else if ($filter.hasClass('filter-boolean')) {
           /** ***********\
            * BOOLEAN
            \************ */
@@ -66,7 +68,6 @@ export default () => {
           cy.scrollTo(0, 0);
           cy.wait(200);
           cy.get('[data-qa="filter-list-chip"]').click({ force: true });
-          cy.get('[data-qa="filter-list-chip"]').click({ force: true });
           cy.get('[data-qa="filter-boolean-false"]').click();
           cy.get('[data-qa="filter-apply"]').click();
           cy.wait(200);
@@ -80,13 +81,14 @@ export default () => {
           });
           cy.scrollTo(0, 0);
           cy.wait(200);
-        } else if ($filter.find('.filter-type-number').length > 0) {
+        } else if ($filter.hasClass('filter-number')) {
           /** ***********\
            * NUMBER
            \************ */
           cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(3);
 
           cy.get('[data-qa="filter-apply"]').click();
+          cy.wait(100);
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -97,12 +99,11 @@ export default () => {
             });
           });
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-number [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('<').click();
 
-          cy.get('[data-qa="filter-apply"]').as('filterApply').click();
+          cy.get('.filter-number [data-qa="filter-apply"]').as('filterApply').click();
+          cy.wait(100);
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -114,12 +115,11 @@ export default () => {
           });
 
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-number [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('<=').click();
 
-          cy.get('[data-qa="filter-apply"]').as('filterApply').click();
+          cy.get('.filter-number [data-qa="filter-apply"]').as('filterApply').click();
+          cy.wait(100);
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -131,12 +131,11 @@ export default () => {
           });
 
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-number [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('>').click();
 
-          cy.get('[data-qa="filter-apply"]').as('filterApply').click();
+          cy.get('.filter-number [data-qa="filter-apply"]').as('filterApply').click();
+          cy.wait(100);
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -148,12 +147,10 @@ export default () => {
           });
 
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-number [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('>=').click();
 
-          cy.get('[data-qa="filter-apply"]').as('filterApply').click();
+          cy.get('.filter-number [data-qa="filter-apply"]').as('filterApply').click();
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -165,8 +162,6 @@ export default () => {
           });
 
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
           cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('between').click();
           cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(2);
@@ -175,7 +170,7 @@ export default () => {
           cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(2);
           cy.get('[data-qa="filter-number-to"]').type('{selectall}').type(6);
           cy.get('[data-qa="filter-apply"]').as('filterApply').click();
-
+          cy.wait(100);
           cy.wait('@apiMemberQuery');
 
           cy.get('@apiMemberQuery').then((req) => {
@@ -186,14 +181,14 @@ export default () => {
               cy.wrap(+value).should('be.lte', 6);
             });
           });
-        } else if ($filter.find('.filter-with-operator-and-input[default-operator="textContains"]').length > 0) {
+        } else if ($filter.hasClass('filter-string')) {
           /** ***********\
            * STRING
            \************ */
-          cy.get('.filter-with-operator-and-input input').clear().type('an');
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('[data-qa="filter-input"]').clear().type('an');
+          cy.get('.filter-string [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('contains').click();
-          cy.get('[data-qa="filter-apply"]').click();
+          cy.get('.filter-string [data-qa="filter-apply"]').click();
           cy.wait(200);
           cy.wait('@apiMemberQuery');
           cy.get('@apiMemberQuery').then((req) => {
@@ -204,12 +199,10 @@ export default () => {
             });
           });
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-string [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('is').click();
 
-          cy.get('[data-qa="filter-apply"]').click();
+          cy.get('.filter-string [data-qa="filter-apply"]').click();
           cy.wait(200);
           cy.wait('@apiMemberQuery');
           cy.get('@apiMemberQuery').then((req) => {
@@ -220,12 +213,10 @@ export default () => {
             });
           });
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
+          cy.get('.filter-string [data-qa="filter-inline-select"]').click();
           cy.get('[data-qa="filter-inline-select-option"]').contains('is not').click();
 
-          cy.get('[data-qa="filter-apply"]').click();
+          cy.get('.filter-string [data-qa="filter-apply"]').click();
           cy.wait(200);
           cy.wait('@apiMemberQuery');
           cy.get('@apiMemberQuery').then((req) => {
@@ -236,19 +227,17 @@ export default () => {
             });
           });
           cy.get('[data-qa="filter-list-chip"]').click();
-          cy.wait(100);
-          cy.get('[data-qa="filter-list-chip"]').click();
-          cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
-          cy.get('[data-qa="filter-inline-select-option"]').contains('not contains').click();
+          cy.get('.filter-string [data-qa="filter-inline-select"]').click();
+          cy.get('[data-qa="filter-inline-select-option"]').contains('does not contain').click();
 
-          cy.get('[data-qa="filter-apply"]').click();
+          cy.get('.filter-string [data-qa="filter-apply"]').click();
           cy.wait(200);
           cy.wait('@apiMemberQuery');
           cy.get('@apiMemberQuery').then((req) => {
             const { rows } = req.response.body;
             rows.forEach((row) => {
               const value = row.attributes[attributeName].default;
-              // cy.wrap(value).should('not.contain', 'an');
+              cy.wrap(value).should('not.contain', 'an');
             });
           });
         } else {
