@@ -21,11 +21,10 @@ export default () => {
     cy.get('[data-qa="filter-apply"]').should('be.disabled');
   });
 
-  // TODO: Enable when bug fixed
-  // it('has apply button disabled if negative value', () => {
-  //   cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(-3);
-  //   cy.get('[data-qa="filter-apply"]').should('be.disabled');
-  // });
+  it('has apply button disabled if negative value', () => {
+    cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(-3);
+    cy.get('[data-qa="filter-apply"]').should('be.disabled');
+  });
 
   it('fetches members with exactly 3 activities', () => {
     cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(3);
@@ -43,8 +42,6 @@ export default () => {
 
   it('fetches members with exactly 3 activities - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(3);
     cy.get('[data-qa="filter-include-switch"]').click();
     cy.get('[data-qa="filter-apply"]').click();
@@ -59,8 +56,6 @@ export default () => {
   });
 
   it('fetches members with less than 3 activities', () => {
-    cy.get('@filterItem').click();
-    cy.wait(100);
     cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('<').click();
@@ -79,8 +74,6 @@ export default () => {
 
   it('fetches members with less than 3 activities - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('<').click();
     cy.get('[data-qa="filter-include-switch"]').click();
@@ -97,8 +90,6 @@ export default () => {
   });
 
   it('fetches members with less than or equal 3 activities', () => {
-    cy.get('@filterItem').click();
-    cy.wait(100);
     cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('<=').click();
@@ -117,8 +108,6 @@ export default () => {
 
   it('fetches members with less than or equal 3 activities - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('<=').click();
     cy.get('[data-qa="filter-include-switch"]').click();
@@ -135,8 +124,6 @@ export default () => {
   });
 
   it('fetches members with more than 3 activities', () => {
-    cy.get('@filterItem').click();
-    cy.wait(100);
     cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('>').click();
@@ -155,8 +142,6 @@ export default () => {
 
   it('fetches members with more than 3 activities - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('>').click();
     cy.get('[data-qa="filter-include-switch"]').click();
@@ -173,8 +158,6 @@ export default () => {
   });
 
   it('fetches members with more than or equal 3 activities', () => {
-    cy.get('@filterItem').click();
-    cy.wait(100);
     cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('>=').click();
@@ -193,8 +176,6 @@ export default () => {
 
   it('fetches members with more than or equal 3 activities - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('>=').click();
     cy.get('[data-qa="filter-include-switch"]').click();
@@ -211,8 +192,6 @@ export default () => {
   });
 
   it('has apply button disabled if range end value is empty', () => {
-    cy.get('@filterItem').click();
-    cy.wait(100);
     cy.get('@filterItem').click();
     cy.get('.filter-list-item-popper [data-qa="filter-inline-select"]').click();
     cy.get('[data-qa="filter-inline-select-option"]').contains('between').click();
@@ -239,8 +218,6 @@ export default () => {
 
   it('fetches members with activity count between 2 and 6 - exclude', () => {
     cy.get('@filterItem').click();
-    cy.wait(100);
-    cy.get('@filterItem').click();
     cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(2);
     cy.get('[data-qa="filter-number-to"]').type('{selectall}').type(6);
     cy.get('[data-qa="filter-include-switch"]').click();
@@ -256,13 +233,10 @@ export default () => {
     });
   });
 
-  // TODO: Enable when bug fixed
-  // it('has apply button disabled if range is invalid', () => {
-  //   cy.get('@filterItem').click();
-  //   cy.wait(100);
-  //   cy.get('@filterItem').click();
-  //   cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(22);
-  //   cy.get('[data-qa="filter-number-to"]').type('{selectall}').type(6);
-  //   cy.get('[data-qa="filter-apply"]').should('be.disabled');
-  // });
+  it('has apply button disabled if range is invalid', () => {
+    cy.get('@filterItem').click();
+    cy.get('[data-qa="filter-number-from"]').type('{selectall}').type(22);
+    cy.get('[data-qa="filter-number-to"]').type('{selectall}').type(6);
+    cy.get('[data-qa="filter-apply"]').should('be.disabled');
+  });
 };
